Update BaseReload API
