+++ uvicorn/supervisors/basereload.py
@@ -1,11 +1,11 @@
 import logging
-import multiprocessing
 import os
 import signal
-import sys
-import time
+import threading
 
 import click
+
+from uvicorn.subprocess import get_subprocess
 
 HANDLED_SIGNALS = (
     signal.SIGINT,  # Unix signal 2. Sent by Ctrl+C.
@@ -20,63 +20,63 @@ class BaseReload:
         self.config = config
         self.target = target
         self.sockets = sockets
-        self.should_exit = False
-        self.reload_count = 0
+        self.should_exit = threading.Event()
+        self.pid = os.getpid()
 
-    def handle_exit(self, sig, frame):
-        self.should_exit = True
-
-    def get_subprocess(self, target, kwargs):
-        spawn = multiprocessing.get_context("spawn")
-        try:
-            fileno = sys.stdin.fileno()
-        except OSError:
-            fileno = None
-
-        return spawn.Process(
-            target=self.start_subprocess, args=(target, fileno), kwargs=kwargs
-        )
-
-    def start_subprocess(self, target, fd_stdin, **kwargs):
-        if fd_stdin is not None:
-            sys.stdin = os.fdopen(fd_stdin)
-        self.config.configure_logging()
-        target(**kwargs)
+    def signal_handler(self, sig, frame):
+        """
+        A signal handler that is registered with the parent process.
+        """
+        self.should_exit.set()
 
     def run(self):
-        pid = str(os.getpid())
+        self.startup()
+        while not self.should_exit.wait(0.25):
+            if self.should_restart():
+                self.restart()
+        self.shutdown()
 
-        message = "Started reloader process [{}]".format(pid)
+    def run(self):
+        self.startup()
+        while not self.should_exit.wait(0.25):
+            if self.should_restart():
+                self.restart()
+        self.shutdown()
+
+    def startup(self):
+        message = "Started reloader process [{}]".format(str(self.pid))
         color_message = "Started reloader process [{}]".format(
-            click.style(pid, fg="cyan", bold=True)
+            click.style(str(self.pid), fg="cyan", bold=True)
         )
         logger.info(message, extra={"color_message": color_message})
 
         for sig in HANDLED_SIGNALS:
-            signal.signal(sig, self.handle_exit)
+            signal.signal(sig, self.signal_handler)
 
-        process = self.get_subprocess(self.target, kwargs=kwargs)
-        process.start()
+        self.process = get_subprocess(
+            config=self.config, target=self.target, sockets=self.sockets
+        )
+        self.process.start()
 
-        while process.is_alive() and not self.should_exit:
-            time.sleep(0.3)
-            if self.should_restart():
-                self.clear()
-                os.kill(process.pid, signal.SIGTERM)
-                process.join()
+    def restart(self):
+        self.mtimes = {}
+        os.kill(self.process.pid, signal.SIGTERM)
+        self.process.join()
 
-                process = self.get_subprocess(self.target, kwargs=kwargs)
-                process.start()
-                self.reload_count += 1
+        self.process = get_subprocess(
+            config=self.config, target=self.target, sockets=self.sockets
+        )
+        self.process.start()
 
-        message = "Stopping reloader process [{}]".format(pid)
+    def shutdown(self):
+        self.process.join()
+        message = "Stopping reloader process [{}]".format(str(self.pid))
         color_message = "Stopping reloader process [{}]".format(
-            click.style(pid, fg="cyan", bold=True)
+            click.style(str(self.pid), fg="cyan", bold=True)
         )
         logger.info(message, extra={"color_message": color_message})
 
-    def clear(self):
-        pass
-
     def should_restart(self):
-        raise NotImplementedError
+        raise NotImplementedError(
+            "Reload strategies should override should_restart()"
+        )
